Retry exceptions should work for subclasses of the retry exceptions too

+++ tomodachi/transport/aws_sns_sqs.py
@@ -162,22 +162,22 @@ class AWSSNSSQSTransport(Invoker):
                 else:
                     kwargs = {}
                     routine = func(*(obj), **kwargs)
-            except (AWSSNSSQSInternalServiceError, AWSSNSSQSInternalServiceErrorException, AWSSNSSQSInternalServiceException) as e:
-                if message_key:
-                    del context['_aws_sns_sqs_received_messages'][message_key]
-                return
             except Exception as e:
+                if issubclass(e.__class__, (AWSSNSSQSInternalServiceError, AWSSNSSQSInternalServiceErrorException, AWSSNSSQSInternalServiceException)):
+                    if message_key:
+                        del context['_aws_sns_sqs_received_messages'][message_key]
+                    return
                 await cls.delete_message(cls, receipt_handle, queue_url, context)
                 raise e
 
             if isinstance(routine, Awaitable):
                 try:
                     return_value = await routine
-                except (AWSSNSSQSInternalServiceError, AWSSNSSQSInternalServiceErrorException, AWSSNSSQSInternalServiceException) as e:
-                    if message_key:
-                        del context['_aws_sns_sqs_received_messages'][message_key]
-                    return
                 except Exception as e:
+                    if issubclass(e.__class__, (AWSSNSSQSInternalServiceError, AWSSNSSQSInternalServiceErrorException, AWSSNSSQSInternalServiceException)):
+                        if message_key:
+                            del context['_aws_sns_sqs_received_messages'][message_key]
+                        return
                     await cls.delete_message(cls, receipt_handle, queue_url, context)
                     raise e
             else:
